# src/main/java/com/cavidanrahmanov/depoti/entity/HeadCategory.java
package com.cavidanrahmanov.depoti.entity;

public enum HeadCategory {
    NƏQLİYYAT,
    DAŞINMAZ_ƏMLAK,
    ELEKTRONİKA,
    MƏİŞƏT_TEXNİKASI,
    EV_VƏ_BAĞ,
    ŞƏXSİ_ƏŞYALAR,
    UŞAQ_ALƏMİ,
    HOBBİ_VƏ_ASUDƏ,
    HEYVANLAR,
    XİDMƏTLƏR_VƏ_BİZNES,
    İŞ_ELANLARI
}